commit-correções

+++ pdv-api/src/main/java/org/example/pdvapi/controllers/ClienteController.java
@@ -32,7 +32,7 @@ public class ClienteController {
             @ApiResponse(responseCode = "404", description = "Cliente não encontrado")})
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> getById(@PathVariable Long id) {
+    public ResponseEntity<Cliente> getById(@PathVariable int id) {
         return ResponseEntity.ok(clienteService.getById(id));
     }
